Track if a repeated field has been deliberately set
Before, if an unset repeated field was ever read, we would make an empty field array for it. As a consequence, .field? and #{field}! methods would indicate that the field had been set, even though it had only been read from.
This was a weird side effect of reading an unset repeated field and has been corrected.